requirements.txt
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.4.2
tensorflow-cpu==2.15.0.post1
protobuf==3.20.3
joblib==1.3.2
requests==2.32.3
yfinance==0.2.40
lxml==5.2.2
beautifulsoup4==4.12.3